[networking] Add standard network headers to libc, remove kernel headers from network apps
