+++ elkscmd/ktcp/tcp_output.c
@@ -15,11 +15,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-
+#include <arpa/inet.h>
 #include "slip.h"
 #include "ip.h"
 #include "tcp.h"
-#include <linuxmt/arpa/inet.h>
 #include "timer.h"
 #include "tcpdev.h"
 #include "netconf.h"
