+++ elkscmd/test/socket/echo/echoclient.c
@@ -1,41 +1,13 @@
 #include <stdio.h>
 #include <unistd.h>
 #include <stdlib.h>
-#ifndef __linux__ 
-#include <linuxmt/socket.h>
-#include <linuxmt/un.h>
-#include <linuxmt/in.h>
-#include "linuxmt/arpa/inet.h"
-#else
 #include <sys/types.h>
 #include <sys/socket.h>
+#include <arpa/inet.h>
 #include <netdb.h>
-#include<string.h>
-#endif
+#include <string.h>
 
 char *socket_path = "/var/uds";
-
-unsigned long int in_aton(const char *str)
-{
-    unsigned long l = 0;
-    unsigned int val;
-    int i;
-
-    for (i = 0; i < 4; i++) {
-	l <<= 8;
-	if (*str != '\0') {
-	    val = 0;
-	    while (*str != '\0' && *str != '.') {
-		val *= 10;
-		val += *str++ - '0';
-	    }
-	    l |= val;
-	    if (*str != '\0')
-		str++;
-	}
-    }
-    return htonl(l);
-}
 
 static void usage()
 {
