+++ libc/include/netdb.h
@@ -1,0 +1,3 @@
+
+/* Absolute file name for network data base files*/
+#define _PATH_HOSTS		"/etc/hosts"
